Another step in that direction despite reservations
